requirements.txt
absl-py==0.13.0
appnope==0.1.2
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
black==21.9b0
bleach==4.1.0
bpemb==0.3.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
configparser==5.0.2
conllu==4.4.1
cramjam==2.3.2
crayons==0.4.0
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.13
entrypoints==0.3
fastparquet==0.7.1
filelock==3.0.12
flair==0.9
flatbuffers==1.12
fsspec==2021.8.1
ftfy==6.0.3
future==0.18.2
gast==0.4.0
gdown==3.12.2
gensim==3.8.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
huggingface-hub==0.0.16
hyperopt==0.2.5
idna==3.2
importlib-metadata==3.10.1
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
Janome==0.4.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
konoha==4.6.5
langdetect==1.0.9
Logbook==1.5.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.8.0
mpld3==0.3
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nltk==3.6.4
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.0
pandas==1.3.2
pandas-datareader==0.10.0
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
plotly==5.3.1
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
segtok==1.5.10
selenium==3.141.0
sentencepiece==0.1.95
six==1.15.0
smart-open==5.2.1
sqlitedict==1.7.0
statsmodels==0.13.0
TA-Lib==0.4.21
tables==3.6.1
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
testpath==0.5.0
textblob==0.15.3
threadpoolctl==2.2.0
thrift==0.13.0
tokenizers==0.10.3
tomli==1.2.1
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
transformers==4.10.0
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
Werkzeug==2.0.1
Wikipedia-API==0.5.4
wrapt==1.12.1
zipp==3.5.0
